refactor: extract shared lead CSV normalization helpers
